feat #584: Ajustes
- Se mejora el funcionamiento de los ajustes en un solo componente
