+++ src/app/pages/ajustes/ajustes.module.ts
@@ -21,7 +21,6 @@ import { ThemeModule } from '../../@theme/theme.module';
 import { ComprobanteComponent } from './comprobante/comprobante.component';
 import { ConsultaAjusteComponent } from './consulta-ajuste/consulta-ajuste.component';
 import { CrudAjusteComponent } from './crud-ajuste/crud-ajuste.component';
-import { CrearAutoComponent } from './crear-auto/crear-auto.component';
 
 @NgModule({
   declarations: [
@@ -30,7 +29,6 @@ import { CrearAutoComponent } from './crear-auto/crear-auto.component';
     ComprobanteComponent,
     ConsultaAjusteComponent,
     CrudAjusteComponent,
-    CrearAutoComponent,
   ],
   imports: [
     CommonModule,
