+++ src/assets/i18n/es.json
@@ -1578,8 +1578,10 @@
       "infoFechaCorte": "Fecha hasta la que se liquidará la amortización"
     },
     "ajuste-auto": {
+      "consultaTtl": "Ajustes automáticos",
+      "consultaSbttl": "A continuación se listan los ajustes generados, seleccione un ajuste para ver el detalle.",
       "registroTtlS": "Resultado Ajuste {{CONSECUTIVO}}",
-      "sbtttlInfo": "A continuación se muestran los elementos ajustados y la respectiva transacción contable correspondiente a este ajuste (si aplica)",
+      "sbtttlInfo": "A continuación se muestran los elementos ajustados y la respectiva transacción contable correspondiente a este ajuste (si aplica).",
       "sugActa": "Seleccione el acta a la que corresponden los elementos que deben ser ajustados",
       "subttlActa1": "Haga click sobre un campo para editarlo.",
       "subttlActa2": "El ajuste se realizará con base en los elementos modificados.",
@@ -1589,8 +1591,7 @@
       "no-necesario-txt": "No se realizaron cambios para generar el ajuste.",
       "alerta-success-ttl": "Ajuste Realizado",
       "alerta-success-inv-txt": "Los elementos fueron actualizados correctamente sin afectaciones contables.",
-      "alerta-success-cont-txt": "Los elementos fueron actualizados correctamente y se generaron transacciones contables.",
-      "volver": "Lista de ajustes"
+      "alerta-success-cont-txt": "Los elementos fueron actualizados correctamente y se generaron transacciones contables."
     }
   }
 }
